iss #96, fetch in Home compnent rather than Event and Quote compnents
