+++ components/home/Home.tsx
@@ -8,23 +8,41 @@ import { auth } from "@/config/Firebase";
 import GoogleLoginBtn from "../utils/GoogleLoginBtn";
 import { useEvent } from "@/context/EventContext";
 import { displayErrorToast, displayToast } from "@/functions/displayToast";
-import LoadingSpinnerL from "../utils/LoadingSpinnerL";
 import { useQuote } from "@/context/QuoteContext";
 import { useAuth } from "@/context/AuthContext";
-import useFetchQuoteFromQuotableAPI from "../hooks/useFetchQuoteFromQuotableAPI";
-import { createProperUrl } from "@/functions/createProperUrl";
 import LoadingIndicator from "./LoadingIndicator";
 
 const Home = () => {
   const [isLoading, setIsLoading] = useState<boolean>(false);
 
   const { loginUser, fetchLoginUser } = useAuth();
+  const { randomEvent, lockedEvent, getRandomEvent, getLockedEvent } =
+    useEvent();
+
+  const { randomQuote, lockedQuote, updateRandomQuote, getLockedQuote } =
+    useQuote();
 
   const [user] = useAuthState(auth);
   useEffect(() => {
+    const fetchEvents = () => {
+      setIsLoading(true);
+      if (!lockedEvent) getLockedEvent();
+      if (auth.currentUser && !randomEvent) getRandomEvent();
+    };
+    const fetchQuotes = async () => {
+      getLockedQuote();
+      updateRandomQuote();
+    };
+
     if (user) {
       try {
         fetchLoginUser(auth.currentUser);
+        if (!(lockedEvent || randomEvent) && !lockedEvent && !randomEvent) {
+          fetchEvents();
+        }
+        if (!(lockedQuote || randomQuote) && !lockedQuote && !randomQuote) {
+          fetchQuotes();
+        }
       } catch (error) {
         displayErrorToast(error);
       } finally {
@@ -39,15 +57,23 @@ const Home = () => {
   }, [user]);
 
   if (isLoading) {
-    return <LoadingIndicator text={"Loading an Login User..."} />;
+    return <LoadingIndicator text={"Loading contents..."} />;
   } else {
     if (!auth.currentUser) {
       return <GoogleLoginBtn />;
     } else {
       return (
         <>
-          <Event />
-          <Quote />
+          {!lockedEvent && !randomEvent ? (
+            <LoadingIndicator text={"Loading an Event..."} />
+          ) : (
+            <Event />
+          )}
+          {!lockedQuote && !randomQuote ? (
+            <LoadingIndicator text={"Loading a Quote..."} />
+          ) : (
+            <Quote />
+          )}
         </>
       );
     }
