+++ components/home/event/Event.tsx
@@ -3,40 +3,13 @@ import { useEvent } from "@/context/EventContext";
 import { auth } from "@/config/Firebase";
 import EventCard from "./EventCard";
 import NoEventAvailable from "./NoEventAvailable";
-import { useEffect, useState } from "react";
-import { displayErrorToast } from "@/functions/displayToast";
-import { useAuth } from "@/context/AuthContext";
-import LoadingSpinnerL from "@/components/utils/LoadingSpinnerL";
-import LoadingIndicator from "../LoadingIndicator";
+import { useAuthState } from "react-firebase-hooks/auth";
 
 const Event = () => {
-  const [isLoading, setIsLoading] = useState<boolean>(true);
+  const [user] = useAuthState(auth);
   const { randomEvent, lockedEvent } = useEvent();
 
-  const { getRandomEvent, getLockedEvent } = useEvent();
-
-  const { loginUser, fetchLoginUser } = useAuth();
-// todo: no need to refetch if exists
-  useEffect(() => {
-    const fetchEvents = () => {
-      getLockedEvent();
-      if (auth.currentUser) getRandomEvent();
-    };
-
-    setIsLoading(true);
-    try {
-      if (loginUser) fetchLoginUser(auth.currentUser);
-      fetchEvents();
-    } catch (error) {
-      displayErrorToast(error);
-    } finally {
-      setTimeout(() => setIsLoading(false), 500);
-    }
-  }, []);
-
-  if (isLoading) {
-    return <LoadingIndicator text={"Loading an Event..."} />;
-  }
+  // todo: no need to refetch if exists
 
   if (!lockedEvent && !randomEvent) {
     return <NoEventAvailable />;
@@ -45,6 +18,7 @@ const Event = () => {
   } else if (randomEvent) {
     return <EventCard event={randomEvent} type="random" />;
   }
+
   return <div>Something wrong here</div>;
 };
